# Upgrade go-ipfix to v0.5.8