see https://github.com/vmware/go-ipfix/releases/tag/v0.5.8 for changes.

With basic structure change of NewInfoElementWithValue, memory usage is expected to be reduced. Here is the Flow Exporter benchmarking improvement for conntrack connections:
Before:
|# conns | iterations | ns/op | B/op | allocs/op|
|-- | -- | -- | -- | --|
|20000   |  100	 |  3484688 ns/op	|  527820 B/op	|    8214 allocs/op|
|30000  |  100  |  5868374 ns/op	|  788098 B/op        |   12313 allocs/op|
|40000  |   100 | 7300047 ns/op     | 1047562 B/op	|   16392 allocs/op|
|50000  |  100	 |  9312464 ns/op	| 1308313 B/op	|   20503 allocs/op|

After change:
|# conns | iterations | ns/op | B/op | allocs/op|
|-- | -- | -- | -- | --|
|20000   |   100	|   3174982 ns/op 	 | 328104 B/op	|    3262 allocs/op|
|30000   |   100	 |  5074667 ns/op	|  489624 B/op  |  4874 allocs/op|
|40000   |  100	|   5910591 ns/op	|  649683 B/op  | 6442 allocs/op|
|50000   |  100	|   8435327 ns/op	|  811341 B/op	  |  8057 allocs/op|

Improvements - reduce by:
|# conns | iterations | ns/op | B/op | allocs/op|
|-- | -- | -- | -- | --|
|20000   |  100	 |  8.89%	|  37.8%	|   60.29%|
|30000  |  100  |  13.53%|  37.87%        |   60.42%|
|40000  |   100 |  19.03%    | 37.98% 	|   60.70%|
|50000  |  100	 |  9.42%	| 37.99%	|   60.70%|

From the result we can see a stable 37% reduction in memory usage and 60% reduction in allocations. Flow Exporter benchmarking of denied connections also achieve around 36% reduction in memory usage and 50% reduction in allocations.