@srikartati 2021-09-24
@zyiou Just remembered about perf test results. As we have some perf fixes in go-ipfix v0.5.8, the perf tests in flow exporter pkg need to be updated especially after Anlan's patch. Could you take care of this?