@zyiou 2021-09-25
Found bandwidth calculation flakiness in this PR and created the fix for it [here](https://github.com/antrea-io/antrea/pull/2841). Other than the fix, this PR is ready to be merged.