@zyiou 2021-09-25
> Found bandwidth calculation flakiness in this PR and created the fix for it here. Other than the fix, this PR is ready to be merged.

Bandwidth flakiness is actually introduced by go-ipfix v0.5.8. Fixed it in this PR instead.